pyproject.toml
[project]
name = "avvai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymupdf>=1.26.7",
    "httpx>=0.28.0",
    "python-dotenv>=1.1.0",
    "toml>=0.10.2",
]
